moving least-squares functions in EB.H and cpp
